Improper Operator Fix
Thanks to the forum post I was able to realize that I wasn’t using the
proper operator for the if statement when checking the boolean values.
I was using “=“ instead of “==“. One assigns and the other is a check
for equals. It’s fixed now and the testing controller now passes all
tests.